requirements.txt
streamlit
pandas
pyarrow
pillow
matplotlib
numpy